Move attrs from code_module to a separate module
